Configure default cores for Travis and add env variable
We add an environment variable because when we automatically generate
the bloop launcher configuration file, modifying how the launcher is
called in `argv` will be more difficult, so we simplify it by sourcing
that configuration from the environment.

We also configure the Travis machines to a default of 2 cores (which is
used in all their machines) because otherwise it would use all of the
available cores in the host machine (not the virtualized environment)
and that can cause Bloop's misbehaviours.