Remove race condition when accessing compiler plugins
The logic responsible for enabling compiler plugin classloader had to
access the file. This logic can be called by any concurrent compiler
process and therefore it needs to be thread-safe. In our previous
version, it wasn't, so if two compiler processes would compete for the
same compiler plugin, one would fail with an exception because only one
zip file can be opened at a time. This would cause the failing process
to disable caching, change the scalac options and force an incremental
recompile.

The new logic has a smart mechanisms to acquire the detection of the
cache with no intereference with other compilation processes.