Populate reporter with client-specific problem metadata
The problem metadata is client-specific because clients (such as BSP)
have state. Therefore, we use the BSP-client-specific data to populate
the reporter that we will use to process events coming from an external
compilation.

We add an extra test case in `ModernDeduplicationSpec` as well as
preparing a new test spec for file watching, which is the place where
this problem was shown. However, the file watching spec will have some
serious changes in next commits, so we just leave a plain simple suite
in this commit instead.