Make sure `socket.close()` works fine
In the previous version of `ipcsocket`, the `close()` implementation was
almost a no-op in both the unix domain socket and named pipe client
sockets. On top of that, the socket's `isClose` method was not correctly
overridden.

By experimentation, we found out that the use of `shutdownInput` and
`shutdownOutput` in unix domain sockets was required. And the windows
named pipe implementation was not fully working
https://github.com/scalacenter/bloop/issues/281

However, that is gone with the new version of ipcsocket that correctly
implements `close` and makes `isClose` work too. The close
implementation now closes the named pipe handles of the streams in
Windows and the input/output streams in *nix systems. It follows exactly
what the pynailgun client does.

So, with this change, doing `socket.close()` is all it suffices in our
test suite to close correctly the sockets and make our bsp test suite
work again.