Finish prototype and avoid race conditions
These changes disallow clients from not obtaining the right last
successful result or deduplicating correctly. In the previous code,
there was a race condition in the following scenario:

1. Client `C1` compiles project `B` that depends on `A`.
1. Client `C2` starts compiling project `A` with no change after client
   `C1` has finished compiling `A` and started compiling `B`.
1. Client `C2` recompiles `A` from scratch even though the compilation
   inputs are the same.

This issue happens because there is no global synchronization between
the clients compiling projects and therefore no mechanism to publish
last successful results. This commit and the previous one do just that:
they add a mechanism to publish result right when a compilation is
deduplicated and obtaining the most recent successful result right when
we're deciding if we should deduplicate.