Invalidate symbols with forked and local javac
Java compilation requires a mechanism to invalidate symbols that does
not depend on deleting class files from the classes directories. The
following commit implements that mechanism for both local and forked
javac compilation.

The mechanism becomes complex when supporting forked javac compilation
since we depend on the javac CLI and we cannot access any public API.
This forces us to implement a clever trick to get the compiler to fail
when a file has been invalidated.