Test cancellation works when BSP client exits impromptu
The following commit makes sure cancellation works if the client exits
suddenly without doing the exit/shutdown handshake at the end of the
connection. This is quite critical for those clients (like VS Code) that
can be restarted and that will close the socket streams. Previously,
there was no guarantee that all the ongoing tasks running previous to
the client exit were cancelled. After this test, we guarantee that's the
case.

There are several changes we make in the repository to make this happen,
some of them enable this use case and some others will enable future
changes in this area:

1. Move ipcsocket sources to the bloop repository to be more free to
   make changes to the socket implementations, which are not complete.
   This commit doesn't introduce any change in the socket
   implementations yet.
1. Copy `LanguageServer` and `LanguageClient` to have more control about
   the way we send and receive messages and to fix several bugs. We have
   done some changes to the server implementation so that we can handle
   sudden close of socket streams from the client side. We still use
   some of the machinery from lsp4s. In the future, there will be more
   changes in this area to make sure, for example, that notifications
   are sent in order.
1. Change the implementation of `BspServer` and use a
   `FoldLeftAsyncConsumer` instead of the previous parallel consumer.
   This will change in future iterations, for now we're using this
   consumer because it has cancellation capabilities that the previous
   `foreachParallelAsync` consumer didn't give us.