Avoid spurious cancellation stalls
To test cancellation inside the compiler we implement waits with macros that simulate busy CPU time. Then, we cancel and hope that the cancelled process will finish before the full waiting time done by the macro calls.

This approach was working just fine before but unfortunately can produce spurious stalls and the compiler might miss the cancellation and not call `advance` if all macro invocations are in one single source file. To avoid that problem, this commit splits the macro invocation calls in independent sources that will force the compiler to check `advance`.