Send out server BSP responses/notifications in order
Synchronize on the output stream that writes out the responses to make
sure that the client receives the BSP server interactions in the same
order they happened. This helps fighting critical violations of the
protocol where `taskFinish` notifications could get before `taskStart`
notifications to the client, or where out-of-order receptions in the
client could produce incorrect user behaviors.