Cancel deduplication if ongoing compilation looks dead for 30s
This PR adds a new system property to configure a disconnection time
from a deduplicated compilation process as well as the machinery
required to implement the cancellation.

When a deduplication takes place and an ongoing compilation has not
registered any kind of progress update for a particular time slot (no
log or reporter action, including scalac progress updates), then Bloop
will disconnect the deduplication, cancel the ongoing compilation and
attempt to schedule another fresh compile. This mechanism is done to
avoid any kind of issues that could occur during compilation.

In all the time of testing this feature, I've never encountered a case
where this is necessary. However, it's a safety mechanism that will
allow us to debug misbehaviours if compilations go awry.

Users can modify the disconnection time via a System property. By
default, 30 seconds need to pass with no progress update before a
disconnection takes place.