Use generic classes for the pipelined traversal
Pipelining was using the client classes directory instead of the generic classes directory, which meant that the IR stores were not put in the right order.

This change also fixes a bug that would show up if a project has several source directories that contain each other. Before this fix, the compiler would fail with repeated definition error.