Synchronize CLI clients to the same build
Bloop does not support concurrent compilation of projects in the same
build across two different CLI clients. It does, however, support it for
one CLI client and as many BSP clients as they connect to the build
server. For this reason, this commit implements a way to prevent two CLI
clients from running operations concurrently in the same build.