Migrate sbt benchmark to bloop repository
We migrate the sbt benchmark to our repository to have a simple way of
testing both sbt and bloop for the same project. To this change we'll
follow up with a guide people can use to run benchmarks for their
projects and see how much real speed does bloop give above sbt.