Persist and aggregate processed repositories
Buildpress now stores the processed repositories in a cache file in the root of the buildpress home. The repository cache serves the purpose of knowing whether a sha in a project has changed or not from one run to another one. If it has, then buildpress deletes the previous directory and clones again the project, checking out the right sha.

The buildpress cache file always has knowledge about all repositories that have been processed by the tool.

If a user wants to force the regeneration of an already cloned project, it can pas the `--generate` flag.