Revalidate classes files compiled by dependent projects
Since the preparation of bloop v1.3.0, class file invalidation no longer
relies on the file system state but instead in a list of populated
invalidated class files that exist during the execution of the bloop
server.

One thing we forgot to do when we prototyped the first implementation
was to think about what would happen to files that were moved from a
project to a dependent project, which is a common thing to do when
refactoring code.

In that scenario, if we move class file `Foo` from `B` to its dependent
`A`, the code would fail to compile because class file invalidation
operates with symbol instead of class files, so Zinc would correctly
invalidate the symbol for `Foo` and add it to the list of invalidated
class symbols in the compiler.

This previous behaviour had the issue that the compiler would not find
the symbol for `Foo` because the symbol was invalidated even though it
was present in the write classes directory of `A`. This commit fixes
that and makes sure we filter out any symbol that is generated from
revious projects.