Handle cancellation of hashing process
Prior to this change, our code was assuming that a process that acquired
the hashing of a classpath entry would always produce a result and, on
the consumer side, we were unconditionally waiting on that result.

We are not sure that's always going to be the case; what if the process
is blocked on IO? what if it never returns? To be on the safe side, we
now support cancellation.  A user cancelling the compilation exactly
while this process happens will fulfill the promise with a cancelled
hash and will force consumers to attempt to reacquire the process to
start the hashing again.