Buffer events while busy correctly
The default implementation of `whileBusyDropEventAndSignal` is buggy for
our use case: we want to buffer all events while the destination is busy
and when it's over send the signal. However, the current implementation
doesn't work this way and if any event happens while the destination is
running it gets added to the buffer count and never flushed to the
destination. We fix this buggy behaviour by creating our own.