Rethink reuse of hashes in classpath hashing
Previously we had enabled cancellation in this part of the code, but I
struggled to find scenarios where the cancellation paths would make a
difference here. Instead, we enable an automatic timeout after 15s that
will complete the downstream hash so that other clients can disconnect
from the hash and re-attempt to hash the entry file again.