Change URL to download coursier
Using the previous URL would download the latest stable version
(1.0.0-M9) instead of whatever version is used in the URL! LoL GitHub.

To fix that error, we instead use another URL that will give us the
coursier artifact associated with the right version we want to download.

I tested this manually and it works.