Add Tomasz as a Bloop contributor [DOCS]
This commit adds Tomasz to the list of contributors in our README.
Tomasz has done a lot of great work around the build tool integrations
and has improved the integration with Metals as well as provided support
in Gitter. @tgodzik you should receive an email inviting you to join the
contributors team; on accepting the invitation, you can merge this pull
request yourself! :smile: