Timeout after default 60s trying to acquire CLI lock
We wait on the CLI lock to be free forever. Use instead a default of 60s
waiting time and if the other client doesn't exit before that, then we
give up and exit with a error status code.

This is a friendly setting for CI instances that have a shared bloop
server and that commnicate with it via CLI clients. If the server is
taken, giving up and not forcing an infinite wait is good.

Users can tweak the default timeout in seconds via
`-Dbloop.cli-lock.seconds-to-timeout`