Redesign build pipelining implementation
There have been lots of changes in this redesign. All in all, what we're
doign is we're adapting to the new zinc rebase changes and cleaning up
the most complicated logic in `CompileGraph`. I am way more comfortable
with the current design because it is conceptually simpler and we're in
complete charge of how we populate signatures to compilers.

This design also has two important advantages:

1. If compile deduplication is enabled and both pipelined and
non-pipelined runs have been request by clients, we will be able to
reuse compilations among these requests.
1. We will not populate symbols from in-memory for those build targets
whose compilation has finished. If that's the case, we prefer loading
symbols from classes directories.