Add support for macros in build pipelining
The previous iteration of build pipelining had support for macros but it
was less sophisticated than what we have now. This new mechanism will
give us better results especially for those projects defining macros in
core projects inside a build but only using them either in tests or in
the last projects to compile in a build.