Delete orphaned bsp client classes directories
Bloop currently deletes classes directories uniquely created for bsp
clients upon a client disconnection (both peaceful and forceful).

However, if the server is unexpectedly killed, this deletion will never
be run and the bsp client classes directories will continue to exist in
the file system.

This commit implements a safe mechanism to delete orphaned client
classes directories. The deletion is spawned in the IO pool every time
there is a successful client connection and it's thread-safe.

This will ensure that these orphaned client classes directories are
wiped out forever.

In addition to that, this commit also adds a mechanism to tell if a
client is still connected or not and lays the foundations to run the
deletion of orphaned internal classes directories.