Clean up orphan internal classes directories
Like bsp client classes directories, internal classes directories can
also become orphan if the server is killed while a compilation is going
on. For this reason, this commit adds support for removing them if they
are found.

The deletion mechanism works as follows:

1. When we load a project for the first time, look into the classes
directories mapped to its analysis file.
2. If it's not inside the root dir that is supposed to host all internal
classes directories, we return, otherwise we continue to step 3.
3. We detect what was the generic classes directory from the format of
the internal classes directory. This way, we can check for the prefix
name of all directories in the root internal classes directories. We
remove all those matched directories that are not our analysis classes
directory, as there can only be one valid internal directory per
project.

Note how this is a safe operation because we only do it the first time
we load a project. Subsequent calls to `load` and `loadAsync` will not
trigger this logic. Therefore, we ensure that we eventually remove all
orphan internal directories.