Swap key and value types of `connectedBspClients` map
I accidentally key'd on the absolute path where I should have key'd on
the client info instead, which is the unique class.

Note that the absolute path is the config dir and we don't yet use it,
but we add it nonetheless because it's likely to be used in the future
and it's better than using `Boolean` as a value or a `newKeySet`.