Catch exceptions when computing the compile bundle
The compile bundle can throw if there is some IO error when, for
example, hashing sources or classpath entries. To protect us from that,
we make sure that we catch any exceptions in its computation and we
translate it to our compile result representation so that the user can
now about it.