Fix memory leak in listing directories
Two crashes due to excessive open file pointers in my macOS machine have
fired a red alarm that have led me to discover a memory leak in the way
we were listing directories. This memory leak was caused because
`Files.list` returns an stream that must be closed, but I have not seen
that before because I was converting that stream to a Scala iterator
via the automatic Java converters and at the end of that iteration
resources were not correctly cleaned up.

Now we make sure that we always close these resources.