Remove socket state from `ConnectionHandle`
`ConnectionHandle` is not a handle when it holds the state of the
connection. Therefore, this commit removes any state from it so that
the client is responsible for opening and closing the socket.

This simplifies reasoning about what does `ConnectionHandle` do.
Furthermore, as connection handle is too generic of a name, we settle
on the `ServerHandle` name which makes it more clear that this handle
is mapped to a server instead of a client connection.