Connect restart from DAP with the listening task
The listening logic before was a while loop that lacked any
communication with the current spawned session. Since most of the times
there should only be one open connection and we cannot process more
while that one is running, there is no use in having an independent task
blocking on `serverSocket.accept()`.

Instead, what we do is that we only have one task listening and then
serving a client. When the DAP client signals that it's interested in
restarting the session, then we recurse (safely, because of Task's
trampolining) and start accepting connections again.