Avoid deletion of unnecessary existing buildpress repos
The following commmit reorganizes the code so that if we find a
repository that is not in the configuration file, we don't delete it
directly but check its current URI and compare it with the one we have
from the buildpress configuration file.

This helps us avoid an excessive deletion of directories that are valid
and whose caches are already populated. This is important whenever we
make changes in the community build.