Document and change `DebugSession` to use underlying logger
This commit applies several changes to `DebugSession` and its callers
so that the debug session logger forwards any event to the underlying
bloop logger instead of throwing an unsupported error.

We need to think carefully what should be the semantics of logger
actions such as `warn`, `debug` or `trace` in the debug session logger
given that we do want to let the client know when things go wrong. I
don't revisit the current assumptions of the design of
`DebugSessionLogger` in this commit but it's worth considering
forwarding as many events as possible to the debug client.