Improve and document better `DebugServer` and `DebugSession`
* Rename `DebugServer` to `StartedDebugServer` to make it clear the
  server has already been started and it's *not* an interface responsible
  for managing and starting a debug server.
* Prevent closing twice in `DebugServer.start`
* Document and change `DebugSession` to use underlying logger
* Change some test methods from `start` to other names so that it's
  clear the server was already started.

Addendum: We need to think carefully what should be the semantics of
logger actions such as `warn`, `debug` or `trace` in the debug session
logger given that we do want to let the client know when things go
wrong. I don't revisit the current assumptions of the design of
`DebugSessionLogger` in this commit but it's worth considering
forwarding as many events as possible to the debug client.