Remove displayWarningToUser when version is unsupported
The intent of this piece of code was good but after thinking about it
and trying it out locally I've realized this is too verbose of a message
and this information should be shown in the Metals doctor instead.

Some of this code will be added back in a next commit in a different
place without the `displayWarningToUser`, which IMO should only be used
when the semanticdb plugin could **not** be resolved for a version we
know it's compatible for. This is truly an scenario that is unlikely to
happen and where good communication with users is key.