Don't ever recompile already compiled classes
Fixes a highly critical problem that caused inconsistent
behavior of incremental compiler and could led to abysmal compile times.

Invalidating a class that was already recompiled once and then letting
javac recompile it again caused javac to not actually recompile
the class and not emit the analysis for that class. And because the
class was already invalidated, its previous analysis was pruned,
hence lost completely. A class without analysis was then considered
removed and triggered many API changes, which invalidated classes
that shouldn't really be invalidated.

Impact of the bug:
- Recompiling far too many Java classes, sometimes
  orders of magnitude more
- Recompiling the same classes again and again
- Running too many rounds of incremental compilation,
  eventually triggering a slow "brute-force" transitive step
- Saving incorrect analysis with many classes missing
- Not recompiling dependent classes on the next recompilation, due
  to broken, incomplete analysis loaded from the previous run

Fixes #986, #987