Fixes unnecessary dependency in sbt extractor
This commit removes a bit of logic that was inherited from the previous
extractor that was not configuration based and that was not spotted
before. This extra line was adding an unncessary compile dependency
between projects when the project instead had another dependency (e.g.
test, it) on that project.