Add incremental build load mechanism
This big commit rethinks how build load works to make it incremental. It
adds more tests and documentation to the appropiate internal APIs to
make this area easier to browse in the future.

The previous logic was working perfectly fine but could incur some
performance overhead because every time there was a change in the server
the build load process would restart and that implies also resolving
semanticdb plugins for all scala versions in a project. That is not a
big problem because most of the times they are cached but the overhead
of that operation is nevertheless too expensive, so the incremental
build load process mitigates that.