Delete both new and read-only paths for invalidated symbols
When a symbol is generated, we removed the rebased read-only path from
the invalidated class files map assuming that an invalidated symbol
used always the read-only classes directory path. However, that's not
always the case -- the paths of an invalidated symbol can also use the
new classes directory and not handling that was an oversight. This
commit just fixes that so that we attempt to both remove the read-only
classes dir path and the write-only classes dir path.

Fixes #972.