Implement `-Xfatal-warnings` manually
Having fatal warnings implemented inside the compiler has several
drawbacks:

1. Bloop receives a failed compilation when there is a warning. A failed
compilation doesn't allow us to cache the work done during an
incremental compiler cycle and the compilation products generated are
thrown away.  This means next time we compile we need to redo the work
of compiling all changed source files when we would only want to
recompile the files that produced a fatal warning.

2. When warnings are produced, Metals doesn't have access to the class
files generated by the failed compilation. This affects negatively the
editing experience in the editor.

To work around these limitations, we remove `-Xfatal-warnings` from the
scalac options passed to the compiler and we implement fatal warnings
ourselves in our reporter implementation.

Our implementation promotes warnings to errors just before we're going
to report them but still represents warnings as warnings in the analysis
file. At the same time, it also treats successful compilations with
fatal warnings as errors at the very end of the compilation pipeline,
after all the compilation passes have been done in a build graph, and
guarantees that it works for clients that have their compilations
deduplicated.

All in all, implementing fatal warnings reproduces the previous behavior
in the compiler but without the inconveniences of handling compilation
results as failures when only fatal warnings are solely responsible for
the failed result.

Our implementation makes sure that the next compile request will
recompile only those files with fatal warnings because we empty the hash
of those sources that contain at least a fatal warning before storing it
in our in-memory store and in disk. As the hashes will be empty, the
next recompilation will always attempt to compile those sources in the
first pass, hence giving the expected behavior.