Improved performance of BloopInvalidatingFileManager
The core of the change is to keep the collection
of invalidated classes always up-to-date by incremental updates,
in order to avoid costly filtering on every read. Additionally some
needless conversions between Array and HashSet were also avoided.

This change was found to speed up incremental
compilation of a 2000+ files Java project by a factor larger than 5.