Avoid unnecessary classpath hash changes
Previous to this patch, no-op compiles could be slower than necessary
because the Zinc compiler would think there's a change in the classpath
that would require to exhaustively look for changes in registered
symbols.

This patch solves this problem in two ways:

1. It better filters out directories from the classpath in all the use
   sites of classpath hashes inside Zinc and Bloop. The logic for this
   change can be found in `BloopLookup` and is shared across the call
   sites.
2. We update no-op previous results with a new compile setup that has
   the most recent classpath hashes to avoid doing the expensive
   classpath change to detect if we should recompile a target or not.

These two changes will improve the worst-case no-op compiles in big
targets and keep expensive classpath changes to a minimum.