Make `bloop server` block to imitate current semantics
When `bloop server` is currently run, the current pynailgun script will
block for the duration of the server execution. This commit refactors
the codebase to imitate the same behavior, which is important for
Systemd and Brew services running `bloop server` in the background. If
we were returning immediately, these services would try to run the start
routines several times per second which would create an unnecessary
overhead on the server.