Support cancellation when hashing sources and classpath
This fix will prevent certain instances of blocking behavior in bloop,
especially during the hashing of classpath entries (which are
concurrently shared by different processes). Upon cancellation of one
the tasks holding the task to hash an entry, the other tasks waiting on
its value will now see that the task was cancelled and fail too.