Shade build modules by default
Modules we're shading:

1. sbt-bloop plugins for 0.13 and 1.0
2. bloopgun
2. launcher

The non-shaded modules will use the suffix `-core` from now on while the
shaded ones will retain the current module names. The default jars for
those modules will be shaded because they are the most common scenario
where they might be used.

The build instrumentation is done in such a way that doing `publish*`
tasks on the respective shaded modules will automatically publish a
shaded module so that they are released automatically with
`releaseEarly` et al.