Reuse previous resolution of configuration files in sbt
In preparation for the offloading compilation feature in sbt, we need to
make sure we never lose the information in the resolutions field as it
is important for the good behavior of Metals.

Given a build that was exported with source information from the Metals
interface, we want sbt-bloop to continue to preserve that information if
it runs `bloopGenerate` on its own before offloading compilation to sbt.