Fix race condition when cancelling source hashing
If we complete the observer, then there's a race condition between
setting `isCancelled` and returning from `mapBoth` and returning the
already hashed sources as a valid result.

To avoid this issue, we don't complete the observer and just wait on the
file discovery to stop when the cancellation is enabled.