Add caching to recursive loop in `Dag.reduce`, fixes #1043.
Previously, a BSP compile request would not respond when compiling a
workspace with a large number of targets. The profile pointed at
`Dag.reduce`, which is a recursive method traversing the dependency
graph between targets.

This commit adds a cache to the recursive loop in `Dag.reduce`, making
the BSP compile requests respond as expected in the same large
workspace.