Recompile if events are watched during compilation
I initially thought this would be more difficult to implement because
of a failed early experiment I did a few weeks ago. After having a
closer look, I've realized that just changing the semantics of
`FoldLeftAsyncConsumer` to wait on the result of the previous task
before running more is enough to make this workflow work.

This is for now quite dumb: any change to a potentially valid Scala/Java
file will retrigger a compilation. In the future, I'd like the
recompilation logic to be smarter and only be executed if the change in
the Scala file wasn't picked up by a compilation process (by comparing
the hashes of the source in the compile bundle and the hash after the
change is received).

Fixes https://github.com/scalacenter/bloop/issues/980