Don't act immediately on MODIFY events with file size
This is a workaround for an issue that happens when using "Remote
Development" in VS Code. When the user runs the "Save file" action in
their remote frontend, the current implementation in the server seems to
cause two side effects:

1. First, it empties out the file with size == 0.
2. Then, it fills in the new contents of the saved file.

One would expect that the save action is atomic, but in this scenario it
isn't. Therefore, sometimes Metals sends a compile request between the
two side effects, meaning it can get to compile the project with an
empty source file which usually generates errors. If the file watcher is
running in the background, it acts on the first side effect and
deduplicates from BSP, the errors will be shown in the terminal. This
workaround here waits 500ms before acting on a MODIFY event for a file
with size == 0, which means the file watcher will not act on the first
side effect but the second. If the second was never to happen after
500ms, then the first side effect would be resubmitted and would
generate a build action.