Add dotty support back to Bloop
https://github.com/scalacenter/bloop/pull/460 initially added support
for Dotty but we removed it recently while experimenting with some
breaking changes in the Zinc compiler interface. Those changes were
afterwards refactored to be non-breaking with the Dotty compiler bridge,
which means Dotty continues to work with Bloop.

However, we must need a change in the Dotty compiler interface to
support invalidation of class files by hooking up into their compiler
symbol loaders. I'll follow up with that change in the Dotty repository.