Implement breakpoint support for user Java/Scala sources
This commit implements breakpoint support for user-provided Java and
Scala sources. It does not support breakpoint support for third-party
sources because those require the intervention of the tool managing them
(i.e. IDEs such as IntelliJ and Metals), see
https://github.com/scalacenter/bloop/issues/1057 for more information.

The way this support works is by reusing the mapping between sources and
class files found in the analysis files. Once we have the analysis files
mapped to a source, we parse them with ASM and extract all the lines in
the `LineNumberTable`, which effectively contain all lines containing
instructions that the user can add breakpoints on. We build a map of
these instruction lines to the fully qualified class name that owns it
and then return back to java-debug, which will do its job.

IDEs that are confident in the reliability of their data can always pass
us the fully qualified class names for some positions themselves via the
URI format "dap-fqcn:$FULLY_QUALIFIED_CLASS_NAME". This is the URI
encoding that should be used to tell Bloop what names should be used for
third-party sources. It is optional and not recommended to pass these
names for user-defined sources and instead let Bloop do its job as its
implementation is as reliable as it can get.

In addition to this, the commit changes some fundamental parts of the
test code to ensure that we can listen for several DAP event
notifications. The previous code only allowed to do `first` and block on
all types of an event. The introduced test changes in this PR allow a
richer, Monix-friendly way of querying the streams of events coming from
the DAP server and block on them only if necessary.