Cache generated project when request comes from same universe id
Sbt can start new universes by extracting state projects and
reevaluating the set of compiled settings. We add a new setting id to
get a proxy for a global unique ID that refers to a concrete
initialization of settings. When a request comes from another universe,
we reevaluate `bloopGenerate` because the new universe could have
changes in the project model that should invalidate the caching.