Add capability to outlive `CTRL-Z` in sbt shell
`CTRL-Z` sends a SIGTSTP signal that corrupts the socket streams. The
issue is less serious with TCP than with Unix Domain sockets, as with
the former the sockets are closed and in the latter one they get blocked
on `read` and never get restarted. Ultimately, compiling, then
suspending the sbt shell and then `fg` + compile breaks either the
previous server connection or blocks it indefinitely.

This commit fixes that by listening to the SIGTSTP signal whenever it
happens, a similar trick to what jline does to handle this very same
scenario. When we detect it, we close the streams of the launcher and
then we start a new connection. Keep in mind that we don't close the
streams from lsp4j (`clientIn` and `clientOut`) because lsp4j logs an
ugly closed stream exception. To work around this, we only close the
launcher streams which has the side-effect of closing the lsp4j session
gracefully.