Support cleaning and cancellation from sbt shell
This commit implements changes in the server to support cleaning from
BSP clients and chagnes in the sbt bloop client to send clean requests
and support cancellation of tasks.

Furthermore, we add some more tests to guarantee that some extra
invariants hold. For example, if the analysis file is wiped out and
there's a no-op compilation, clients might expect that the analysis file
gets repopulated.

Support for cleaning isn't done yet, as we still need to ensure that we
imitate sbt cleaning semantics 100% in a follow-up PR.