Imitate same sbt clean semantics
The `clean` task in sbt wipes out everything under `target`, which means
it affects all compilation outputs of all the configurations. To imitate
the same behavior, running `clean` scoped to the project sends a clean
to the server with all the build target ids mapped to the existing
configurations.

Furthermore, sbt doesn't do anything when the `clean` task is invoked
scoped to a concrete configuration. Unlike sbt, we decide to send a
clean request so that users can work around against this annoying
default behavior.