Add the build machinery to work around sbt metaproject bug
See details in https://github.com/sbt/sbt/issues/892. We were getting
same error because the plugin comes from a jar. Now, we lie to sbt, we
present the plugin as a project with classes directory which just
happens to have a dependency that brings in the shaded sbt plugin we
want.