Make sbt-bloop transform `Provided` to `Compile` deps
Before, when a project had a reference to a `Provided` dependency in
either the origin or target dependency configuration, Bloop would skip
it. However, the only thing that makes sense to do in these scenarios is
to pick use the `Compile` configuration so that provided dependencies
are considered compilation dependencies in Bloop. This is what sbt
ultimately does by adding targets of provided projects into the internal
compile classpath.

Without this change, some projects would fail to compile because the
target directories of their classpath wouldn't be populated. With this
change, the projects will always succeed to compile because before
compiling them Bloop will compile the dependencies that are provided.