Refactor `JavaEnv` into `JdkConfig` and `JavaRuntime`
The previous concept of `JavaEnv` didn't make a lot of sense and was a
combination of many jdk-related/java-related values and methods.

Instead, we clearly refactor that into two different concepts.
`JdkConfig` takes care of encapsulating the information required to
configure the JDK, either for compiling (`javac`) or for
running/testing. `JavaRuntime` is a utility that contains values and
methods related to the current Java runtime.