Add code to infer name from closeness to line number
This code is not necessary and should not be run if everything works as
expected as specified by the javac compiler and the JDI specification.

However, when there's not a line number match, we don't want to fail to
infer the class name where the line is defined. Therefore, we do a
best-effort heuristic so that whenever there's a code line miss we try
to find the instruction line number that is closes to it (both from
above and below) and select the class file name that owns that
instruction.

If that instruction does indeed happens to be invalid and that's why we
run this code, then the java-debug implementation will mark it as
invalid and not verify it, which is what we want from a user point of
view. This code only adds an extra layer of protection in case
something goes wrong on our side.