Drop dependency between release and test jobs.
Previously, the release job only ran if tests passed. This caused the
release to not run on flaky test failures. Now, the release job runs
regardless if the tests pass or not.

A nice effect of this change is that releases finish faster. We use the
same setup in Metals and have not experienced issues with releasing
unwanted commits. If something lands into master we want to release it
regardless of test results.