Require collection field types in jsoniter codecs
Previously, Scala collection types in generated jsoniter codecs could be
left empty and the deserialization would not fail. However, this runs
constrary to the semantics in our previous Circe codecs and the
semantics we want to enforce: all fields should be mandatory unless they
are explicitly wrapped on an `Option` type.

This PR enables a new option that @plokhotnyuk introduced in https://github.com/plokhotnyuk/jsoniter-scala/pull/411/files